requirements.txt
Pillow == 8.0.1
numpy == 1.15.0